% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_titles.R
\docType{data}
\name{area_titles}
\alias{area_titles}
\title{Dataset containing FIPS codes for counties, states and MSAs.}
\format{A data frame with 4723 rows and 2 variables}
\usage{
data(area_titles)
}
\description{
Built-in dataset to look up areas by FIPS code.
Includes recent and historical FIPS codes.
To access the data directly, issue the command \code{area_titles)}.
\itemize{
  \item \code{area_fips}: FIPS code
  \item \code{area_title}: Area name
}
}
\note{
Last updated 2016-09-30
}
\keyword{internal}
