% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bls_key}
\alias{bls_key}
\title{Utility to store your BLS API key.}
\usage{
bls_key(force = FALSE)
}
\arguments{
\item{force}{TRUE or FALSE. The default value is FALSE.}
}
\description{
Stores your BLS API key in your .renviron file for security.
}
\examples{

## Not run:
bls_key()

## End (Not run)

}

