% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{bls_map}
\alias{bls_map}
\title{Cloropleth mapping of BLS data.}
\usage{
bls_map(map_data, fill_rate)
}
\arguments{
\item{map_data}{Dataframe to be used as the map's measures. Usually a result of function calls format_county_data or format_state_data, but other dataframes, which include FIPS codes may be used as well.}

\item{fill_rate}{Column name from the dataframe that you want to use as a fill value.}
}
\description{
Return a ggplot object to render a cloropleth map.
}
\examples{

## Not run:
df <- format_county_data()
bls_gg <- bls_map(map_data = df, fill_rate = "unemp_rate")
bls_gg

## End (Not run)


}
\keyword{api}
\keyword{bls}
\keyword{economics}
\keyword{geo}
\keyword{geography}
\keyword{map}
\keyword{unemployment}

