% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{bls_map_state}
\alias{bls_map_state}
\title{Cloropleth mapping of BLS data.}
\usage{
bls_map_state(map_data, fill_rate)
}
\arguments{
\item{map_data}{Dataframe to be used as the map's measures. Usually a result of function calls format_county_data or format_state_data, but other dataframes, which include FIPS codes may be used as well.}

\item{fill_rate}{Column name from the dataframe that you want to use as a fill value.}
}
\description{
Return a ggplot object to render a cloropleth map with state outlines.
}
\examples{

## Not run:
df <- bls_state_data(seasonality = TRUE)

#Subset data frame to most recent month.
recent <- max(df$month)
df <- df[ which(df$month==recent), ]

bls_gg <- bls_map_state(map_data = df, fill_rate = "unemployed_pct")
bls_gg

## End (Not run)


}
\keyword{api}
\keyword{bls}
\keyword{economics}
\keyword{geo}
\keyword{geography}
\keyword{map}
\keyword{unemployment}

