% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_state_data.R
\name{bls_state_data}
\alias{bls_state_data}
\title{Gather State Employment Data from the BLS.}
\usage{
bls_state_data(seasonality = NA)
}
\arguments{
\item{seasonality}{TRUE or FALSE. Would you like seasonally adjusted data? The default value is TRUE.}
}
\description{
Returns data from 1976 to current in a dataframe. Data includes, Civilian Labor Force, Number of Employed, Number of Unemployed and Unemployment rate.
}
\examples{

## Not run:
df <- bls_state_data(seasonality = TRUE)

## End (Not run)


}
\keyword{bls}
\keyword{data}
\keyword{employment,}
\keyword{force,}
\keyword{labor}
\keyword{unemployment,}

