% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Compound_pipe}
\usage{
lhs \%<>\% rhs
}
\value{
No return value, a utility function to export magrittr pipes to be used in internal package code.
}
\description{
Compound_pipe
}
\keyword{internal}
