% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_fips.R
\docType{data}
\name{county_fips}
\alias{county_fips}
\title{Dataset containing FIPS codes for mapping and lookup.}
\format{A data frame with 3085 rows and 3 variables}
\usage{
data(county_fips)
}
\description{
Built-in dataset for use with the \code{bls_map} function.
             To access the data directly, issue the command \code{data(datacounty_fips)}.

\itemize{
  \item \code{fips_county}: Combined FIPS code for county + state
  \item \code{state}: State name
  \item \code{county}: County name
}
}
\details{
Dataset containing FIPS codes for mapping and lookup.

Built-in dataset for use with the \code{bls_map} function.
To access the data directly, issue the command \code{datacounty_fips)}.
}
\note{
Last updated 2016-05-27
}
\keyword{datasets}

