% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cu_main}
\alias{cu_main}
\title{Dataset containing All items in U.S. city average, all urban consumers, seasonally adjusted CUSR0000SA0.}
\format{
A data frame with 926 rows and 4 variables
}
\usage{
data(cu_main)
}
\description{
Built-in dataset to look CPI values by month.
Includes recent and historical CPI values.
To access the data directly, issue the command \code{cu_main)}.
\itemize{
  \item \code{date}: Date in yyyy-mm-dd format
  \item \code{period}: Monthly period
  \item \code{year}: CPI year
  \item \code{value}: CPI value
}
}
\note{
Last updated 2024-03-24
}
\keyword{internal}
