% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstupper}
\alias{firstupper}
\title{Internal function to cast the first letter of a word to upper-case.}
\usage{
firstupper(x)
}
\arguments{
\item{x}{A word or string to capatalize}

\item{...}{additional arguments.}
}
\value{
A string with the first letter of \code{x} as capital letter.
}
\description{
Internal function to cast the first letter of a word to upper-case.
}
\keyword{internal}
