% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_state_data}
\alias{format_state_data}
\title{Helper funtion for state map.}
\usage{
format_state_data(state, seasonality = NA)
}
\arguments{
\item{state}{A list of states to run through the loop.}

\item{seasonality}{TRUE or FALSE. The default value is TRUE.}
}
\description{
Helper function to download and format state employment data.
}

