% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bls_county.R
\name{get_bls_county}
\alias{get_bls_county}
\title{Function that returns county-level labor statistics for the last three months.}
\usage{
get_bls_county()
}
\description{
Helper function to download and format state employment data. Note: This returns only non-seasonally adjusted data.
}

