% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blscrapeR.R
\name{get_data}
\alias{get_data}
\title{Basic Request Mechanism for BLS Tables}
\usage{
get_data(seriesid, startyear = NULL, endyear = NULL,
  registrationKey = NULL, catalog = NULL, calculations = NULL,
  annualaverage = NULL)
}
\arguments{
\item{bls_id}{a specific BLS series id or a list of series ids to be called.}

\item{api.version}{An integer of the BLS veriosn number (1 or 2), 1 is the default, version 2 requires registration.}
}
\description{
Return data frame from one or more requests via the US Bureau of Labor Statistics API. Provided arguments are in the form of BLS series ids.
}
\examples{

## Not run:
## API Version 1.0 R Script Sample Code
## Single Series request
df <- get_data('LAUCN040010000000005')

## End (Not run)


}
\keyword{api}
\keyword{bls}
\keyword{cpi}
\keyword{economics}
\keyword{inflation}
\keyword{unemployment}

