% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics}
\alias{naics}
\title{Dataset containing NIACS codes for industry lookups.}
\format{
A data frame with 2469 rows and 2 variables
}
\usage{
data(naics)
}
\description{
Built-in dataset to look up industries by NAICS code.
To access the data directly, issue the command \code{naics)}.
\itemize{
  \item \code{industry_code}: Industry code
  \item \code{industry_title}: Industry title
}
}
\note{
Last updated 2019-12-15
}
\keyword{internal}
