% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
No return value, a utility function to export magrittr pipes to be used in internal package code.
}
\description{
Pipe operator
}
\keyword{internal}
