% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_functions.R
\name{quick_employed_level}
\alias{quick_employed_level}
\title{Quick employed level}
\usage{
quick_employed_level()
}
\description{
Returns the employment level. SeriesID: LNS12000000
If you installed a BLS_KEY with the \code{set_bls_key()} function, it will detect it and use your key. This counts against your daily query limit.
}
\examples{
\dontrun{
df <- quick_employed_level()
}

}
\keyword{quick}
\keyword{rate}
\keyword{unemployment}
