% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_functions.R
\name{quick_laborForce_level}
\alias{quick_laborForce_level}
\title{Quick Civilian Labor Force Level}
\usage{
quick_laborForce_level()
}
\description{
Returns the civilian labor force level. SeriesID: LNS11000000.
If you installed a BLS_KEY with the \code{set_bls_key()} function, it will dectect it and use your key. This counts against your daily query limit.
}
\examples{
\dontrun{
df <- quick_laborForce_level()
}

}
\keyword{quick}
\keyword{rate}
\keyword{unemployment}
