% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_functions.R
\name{quick_laborForce_rate}
\alias{quick_laborForce_rate}
\title{Quick Civilian Labor Force Rate}
\usage{
quick_laborForce_rate()
}
\value{
A tibble from the BLS API.
}
\description{
Returns the civilian labor force participation rate. SeriesID: LNS11300000.
If you installed a BLS_KEY with the \code{set_bls_key()} function, it will detect it and use your key. This counts against your daily query limit.
}
\examples{
\dontrun{
df <- quick_laborForce_rate()
}

}
\keyword{quick}
\keyword{rate}
\keyword{unemployment}
