% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_functions.R
\name{quick_nonfarm_employed}
\alias{quick_nonfarm_employed}
\title{Quick total nonfarm employment}
\usage{
quick_nonfarm_employed()
}
\value{
A tibble from the BLS API.
}
\description{
Returns the Total Nonfarm Payroll Employment, seasonally adjusted. BLS id CES0000000001.
If you installed a BLS_KEY with the \code{set_bls_key()} function, it will detect it and use your key. This counts against your daily query limit.
}
\examples{
\dontrun{
df <- quick_nonfarm_employed()
}

}
\keyword{nonfarm_employed}
\keyword{quick}
