% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ids.R
\name{search_ids}
\alias{search_ids}
\title{Search the internal series_id data set.}
\usage{
search_ids(keyword = NULL, periodicity_code = NULL, ...)
}
\arguments{
\item{keyword}{The keyword you want to search. This can be a fuzzy search of multiple keywords. For example \code{"unemployment women"}.}

\item{periodicity_code}{The period of time of the surveys you are interested in. This is usually monthly, quarterly or annually.
You can type full words or beginning letters. For example, \code{periodicity_code = "m"} or \code{periodicity_code = "monthly"}.}

\item{...}{additional arguments}
}
\description{
Search the internal series_id data set.
}
\examples{
# Search for monthly Unemployment Rates for Women
ids <- search_ids(keyword = c("Unemployment Rate", "Women"), periodicity_code = "M")

}
