% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_titles.R
\docType{data}
\name{size_titles}
\alias{size_titles}
\title{Dataset containing size codes for US industries by size.}
\format{A data frame with 10 rows and 2 variables}
\usage{
data(size_titles)
}
\description{
Built-in dataset to look up industries by size.
To access the data directly, issue the command \code{size_titles)}.
\itemize{
  \item \code{size_code}: Size code
  \item \code{size_title}: Size title
}
}
\note{
Last updated 2016-09-30
}
\keyword{internal}

