% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fips.R
\docType{data}
\name{state_fips}
\alias{state_fips}
\title{Dataset containing FIPS codes for mapping and lookup.}
\format{A data frame with 63 rows and 6 variables}
\usage{
data(state_fips)
}
\description{
Built-in dataset for use with the \code{bls_map} function.
             To access the data directly, issue the command \code{data(datastate_fips)}.

\itemize{
  \item \code{fips_state}: FIPS code for state
  \item \code{saa}: saa cross-walk id
  \item \code{region}: Geographic region
  \item \code{division}: Geographic division
  \item \code{state_abb}: State abbreviation
  \item \code{state}: State name
}
}
\details{
Dataset with the lat. / long. of county FIPS codes used for mapping.

Built-in dataset for use with the \code{bls_map} function.
To access the data directly, issue the command \code{datastate_fips)}.
}
\note{
Last updated 2016-05-27
}
\keyword{datasets}

