% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_map.R
\docType{data}
\name{state_map_data}
\alias{state_map_data}
\title{Dataset with the lat. / long. of county FIPS codes used for mapping.}
\format{A data frame with 206,500 rows and 7 variables}
\usage{
data(state_map_data)
}
\description{
Built-in dataset for use with the \code{bls_map_state} function.
To access the data directly, issue the command \code{datastate_map_data)}.
}
\details{
\itemize{
  \item \code{long}: State longitude
  \item \code{lat}: State latitude
  \item \code{order}: Polygon order
  \item \code{hole}: hole
  \item \code{piece}: Polygon piece
  \item \code{id}: FIPS Code
  \item \code{group}: group
}
}
\note{
Last updated 2016-05-26
}
\keyword{internal}

