% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urlExists}
\alias{urlExists}
\title{urlExists}
\usage{
urlExists(target)
}
\arguments{
\item{target}{url}
}
\value{
A logical of \code{TRUE} or \code{FALSE}.
}
\description{
A utility function to run a tryCatch on a URL.
}
