% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{bls_map}
\alias{bls_map}
\title{Cloropleth mapping of BLS data.}
\usage{
bls_map()
}
\description{
Return a ggplot object to render a cloropleth map.
}
\examples{

## Not run:
## One or More Series with Optional Parameters
bls_id <- list(
    'seriesid'=c('LNS14000000','LNS12300000'),
    'startyear'=2010,
    'endyear'=2012,
    'catalog'=FALSE,
    'calculations'=TRUE,
    'annualaverage'=TRUE,
    'registrationKey'='2a8526b8746f4889966f64957c56b8fd')
df <- get_data(bls_id, api.version=2)

## End (Not run)


}
\keyword{api}
\keyword{bls}
\keyword{economics}
\keyword{geo}
\keyword{geography}
\keyword{map}
\keyword{unemployment}

