% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{bls_map}
\alias{bls_map}
\title{Cloropleth mapping of BLS data.}
\usage{
bls_map()
}
\description{
Return a ggplot object to render a cloropleth map.
}
\examples{

## Not run:

bls_gg <- bls_map()
bls_gg

## End (Not run)


}
\keyword{api}
\keyword{bls}
\keyword{economics}
\keyword{geo}
\keyword{geography}
\keyword{map}
\keyword{unemployment}

