% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bls_state.R
\name{bls_state_data}
\alias{bls_state_data}
\title{Function that returns county-level labor statistics for the last three months.}
\usage{
bls_state_data(state, seasonality = NA)
}
\arguments{
\item{state}{A list of states to run through the loop.}

\item{seasonality}{TRUE or FALSE. The default value is TRUE.}
}
\description{
Helper function to download and format state employment data.
}

