% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_map.R, R/county_map_merc.R
\docType{data}
\name{county_map_data}
\alias{county_map_data}
\alias{county_map_data}
\alias{county_map_merc}
\title{Dataset for mapping U.S. counties}
\format{A data frame with 55,413 rows and 7 variables}
\usage{
data(county_map_data)

data(county_map_merc)
}
\description{
A fortified data set that includes U.S. counties and is suitable for making maps with \code{ggplot2}.
The county FIPS codes and boundary lines from the most recent TIGER release from the U.S. Census Bureau.

\itemize{
  \item \code{long}: County longitude
  \item \code{lat}: County latitude
  \item \code{order}: Polygon order
  \item \code{hole}: hole
  \item \code{piece}: Polygon piece
  \item \code{id}: FIPS Code
  \item \code{group}: group
}

A fortified data set that includes U.S. counties and is suitable for making maps with \code{ggplot2}.
The county FIPS codes and boundary lines from the most recent TIGER release from the U.S. Census Bureau.

\itemize{
  \item \code{long}: County longitude
  \item \code{lat}: County latitude
  \item \code{order}: Polygon order
  \item \code{hole}: hole
  \item \code{piece}: Polygon piece
  \item \code{id}: FIPS Code
  \item \code{group}: group
}
}
\details{
Dataset with the lat. / long. of county FIPS codes used for mapping.

Built-in dataset for use with the \code{bls_map_county()} function.
To access the data directly, issue the command \code{data(county_map_data)}.

Dataset with the lat. / long. of county FIPS codes used for mapping.

Built-in dataset for use with the \code{bls_map_county()} function.
To access the data directly, issue the command \code{data(county_map_merc)}.
}
\note{
Last updated 2017-01-26

Last updated 2017-01-26
}
\keyword{datasets}
