% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bls_state.R
\name{get_bls_state}
\alias{get_bls_state}
\title{Function that returns county-level labor statistics for the last three months.}
\usage{
get_bls_state(date_mth, seasonality = NA)
}
\arguments{
\item{date_mth}{The month or months you would like data for. Accepts full month names and four-digit year.}

\item{seasonality}{TRUE or FALSE. The default value is TRUE.}
}
\description{
Helper function to download and format state employment data.
}
\examples{
 \dontrun{
# Single series
get_bls_state(date_mth = "May 2016", seasonality = TRUE)

# Multiple series
get_bls_state(date_mth = c("April 2016", "May 2016"), seasonality = FALSE)
}

}

