% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blscrapeR.R
\name{get_data}
\alias{get_data}
\title{Basic Request Mechanism for BLS Tables}
\usage{
get_data(bls_id = NA, api.version = 1)
}
\arguments{
\item{bls_id}{a specific BLS series id or a list of series ids to be called.}

\item{api.version}{An integer of the BLS veriosn number (1 or 2), 1 is the default, version 2 requires registration.}
}
\description{
Return data frame from one or more requests via the US Bureau of Labor Statistics API. Provided arguments are in the form of BLS series ids.
}
\examples{

## Not run:
## API Version 1.0 R Script Sample Code
## Single Series request
df <- get_data('LNS14000000')

## End (Not run)

## Not run:
## Multiple Series
bls_id <- list('seriesid'=c('LNS12300000','LNS14000000'))
df <- get_data(bls_id)

## End (Not run)

## Not run:
## One or More Series, Specifying Years
bls_id <- list(
    'seriesid'=c('LNS14000000','LNS12300000'),
    'startyear'=2010,
    'endyear'=2012)
df <- get_data(bls_id)

## End (Not run)

## Not run:
## API Version 2.0 R Script Sample Code
## Single Series
df <- get_data('LNS14000000', api.version=2)

## End (Not run)

## Not run:
## Multiple Series
bls_id <- list('seriesid'=c('LNS14000000','LNS12300000'))
df <- get_data(bls_id, 2)

## End (Not run)

## Not run:
## One or More Series with Optional Parameters
bls_id <- list(
    'seriesid'=c('LNS14000000','LNS12300000'),
    'startyear'=2010,
    'endyear'=2012,
    'catalog'=FALSE,
    'calculations'=TRUE,
    'annualaverage'=TRUE,
    'registrationKey'='2a8526b8746f4889966f64957c56b8fd')
df <- get_data(bls_id, api.version=2)

## End (Not run)


}
\keyword{api}
\keyword{bls}
\keyword{cpi}
\keyword{economics}
\keyword{inflation}
\keyword{unemployment}

