% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niacs.R
\docType{data}
\name{niacs}
\alias{niacs}
\title{Dataset containing NIACS codes for industry lookups.}
\format{A data frame with 2469 rows and 2 variables}
\usage{
data(niacs)
}
\description{
Built-in dataset to look up industries by NIACS code.
To access the data directly, issue the command \code{niacs)}.
\itemize{
  \item \code{industry_code}: Industry code
  \item \code{industry_title}: Industry title
}
}
\note{
Last updated 2016-09-30
}
\keyword{internal}

