% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{series_ids}
\alias{series_ids}
\title{Dataset containing BLS series ids and descriptions.}
\format{An object of class \code{data.frame} with 79029 rows and 4 columns.}
\usage{
data(series_ids)
}
\description{
Built-in dataset to look up BLS series ids and descriptions.
Currently limited to the CPS and LAUS.
To access the data directly, issue the command \code{series_ids)}.
}
\note{
Last updated 2017-11-15
}
\keyword{internal}
