% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_fips}
\alias{state_fips}
\title{Dataset with the lat. / long. of county FIPS codes used for mapping.}
\format{
A data frame with 57 rows and 4 variables
}
\usage{
data(state_fips)
}
\description{
Built-in dataset for use with the \code{bls_map} function.
To access the data directly, issue the command \code{datastate_fips)}.


\itemize{
  \item \code{fips_state}: FIPS code for state
  \item \code{state_abb}: State abbreviation
  \item \code{state}: State name
  \item \code{gnisid}: Geographic Names Information System ID
}
}
\note{
Last updated 2016-05-27
}
\keyword{internal}
